+++ JBSnorro/JS/JSProcessRunner.cs
@@ -1,8 +1,7 @@
-﻿#pragma warning disable CS1066
+#pragma warning disable CS1066
 using JBSnorro;
 using JBSnorro.Csx;
 using JBSnorro.Csx.Node;
-using JBSnorro.Diagnostics;
 using JBSnorro.Extensions;
 using JBSnorro.IO;
 using JBSnorro.Text;
@@ -412,18 +411,5 @@ const reviver = function (key, value) {
         }
 
         public override string Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options) => throw new NotImplementedException();
-    }
-}
-[DebuggerDisplay("JSSourceCode({Value})")]
-public class JSSourceCode
-{
-    public static JSSourceCode Null { get; } = new JSSourceCode("null");
-    public static JSSourceCode Undefined { get; } = new JSSourceCode("undefined");
-
-    public string Value { get; }
-    public JSSourceCode(string sourceCode)
-    {
-        Contract.Requires(sourceCode != null, nameof(sourceCode));
-        Value = sourceCode;
     }
 }
